Add reproducer script for rabbit problems
there's a heisenbug around where nova-compute sometimes fails
to re-connect to rabbitmq and is left in a stuck state.

The script added is a helper to reproduce this bug.
It restarts rabbitmq and checks state of nova-compute service,
comparing last time state was updated.
If the Updated-At fields stays unchanged for some time, the compute
is obviously stuck.

Kudos to vsaienk0.